refactor: 화면모드전환 ActionSheet 메서드 분리
